Rename: Add.tsx에서 각 page들만 routing하는 형식으로 폴더 및 파일 변환
