+++ client/src/components/daywords/Day.tsx
@@ -1,5 +1,5 @@
 import { useParams } from 'react-router-dom';
-import useFetch from '../hooks/useFetch';
+import useFetch from '../../hooks/useFetch';
 import Word, { IWord } from './Word';
 
 const Day = () => {
