+++ client/src/components/daywords/Word.tsx
