+++ client/src/pages/CreateWordPage.tsx
@@ -1,0 +1,11 @@
+import CreateWord from 'components/createword/CreateWord';
+
+const CreateWordPage = () => {
+  return (
+    <>
+      <CreateWord />
+    </>
+  );
+};
+
+export default CreateWordPage;
