+++ client/src/pages/DayWordsPage.tsx
@@ -1,0 +1,11 @@
+import Day from 'components/daywords/Day';
+
+const DayWordsPage = () => {
+  return (
+    <>
+      <Day />
+    </>
+  );
+};
+
+export default DayWordsPage;
