+++ client/src/pages/HomePage.tsx
@@ -1,0 +1,11 @@
+import DayList from 'components/home/DayList';
+
+const HomePage = () => {
+  return (
+    <>
+      <DayList />
+    </>
+  );
+};
+
+export default HomePage;
