--- client/src/pages/NotFound.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const NotFound = () => {
-  return <div>Not Found</div>;
-};
-
-export default NotFound;
