+++ client/src/pages/NotFoundPage.tsx
@@ -1,0 +1,5 @@
+const NotFoundPage = () => {
+  return <div>Not Found</div>;
+};
+
+export default NotFoundPage;
