+++ client/src/types/days.d.ts
@@ -1,0 +1,6 @@
+interface IDays {
+  id: number;
+  day: number;
+}
+
+export { IDays };
